+++ NorthernSkyRaidTools/Profiles.lua
@@ -196,6 +196,7 @@ function NSI:AddMissingDefaults()
 
         Alerts = {
             ReloeReminders = false,
+            Language = "Auto",
             Groups = {},
         },
 
@@ -566,6 +567,7 @@ local ignored = {
     ["CurrentProfile"]   = true,
     ["MainProfile"]      = true,
     ["EncounterAlerts"]  = true,
+    ["EncounterAlertOriginals"] = true,
     ["AuraTrackingSettings"] = true,
     ["AuraSounds"]       = true,
     ["NickNames"]        = true,